nixos/k3s: don't set cluster DNS in hosts resolv.conf
The hosts' resolv.conf is passed to the k3s cluster coredns, which leads
to coredns trying to _sometimes_ resolve requests to itself.
This leads to timeouts and otherwise failing DNS request.

This change also modifies the behavior of DNS resolving on the host
directly. Especially HAProxy now requires to especially set the CoreDNS
as resolver. We checked that this won't affect customers currently.

PL-134112