telegraf net input: disable legacy protocol metrics
Collection network protocol stats via the `net_` plugins has been
deprecated in telegraf, disabled by default, and even the opt-in
parameter for keeping that behaviour during a transitioning period will
be disabled in telegraf-1.37.
We fully disable the legacy metrics for the upcoming platform release
and provide clear guidance in the upgrade notes to enable customers to
swithc to the new metrics.

PL-133683