skvaider: refactor 25.11 fetching from versions.nix apporach to overlay
Fetching via versions.nix has multiple problems:
- versions.json gets overriden by our nixpkgs update script, if the input
  is not referenced in the flake.nix (bit awkward and error-prone design,
  I know.)
- Adding a new input to versions.json leads to a size increase of the
  channel tarball. For a new nixpkgs version, this is relevant.

This change, changes this to be included from overlay.nix, which is our
central point for non-upstream changes related to packages. This allows
us to discover more easily that this behavior exists and therefore also
to drop it in the future when it's unused.

With this approach, the Nixpkgs checkout is a FOD derivation and only
gets downloaded when the package (here: ollama-rocm) gets used.
This works for all packages, but not for NixOS modules.