doc: update redis docs regarding password changes
This also removes the old way of setting the redis password with a Nix
expression, because that leads to the password being stored
world-readable. This is still possible, but we don't want to advertise
this.

PL-134135

(cherry picked from commit 4da3659ad456a9e1a6022f4376a0404b7b664f34)